# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaspi Pay POS Automation API",
    "version": "1.0.0",
    "description": "REST API for integrating POS systems with Kaspi Pay API (SMS Authorization, Invoicing, QR Payments, Operation History, and Refunds)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Theater API Key used for authentication and permission gating"
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master admin secret token check (ADMIN_SECRET_KEY)"
      },
      "TokenSN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-SN",
        "description": "Session token SN obtained after OTP verification"
      },
      "VtokenSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Vtoken-Secret",
        "description": "Encrypted session token secret"
      },
      "ProfileId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Profile-Id",
        "description": "Organization profile ID (optional, sent after auth)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Description of the error"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/keys": {
      "post": {
        "tags": ["Admin"],
        "summary": "Create Theater API Key",
        "description": "Generates a new API Key for a theater with configured path permissions.",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Theater name",
                    "example": "Grand Cinema"
                  },
                  "allow_invoice": {
                    "type": "boolean",
                    "description": "Enable invoicing permission",
                    "example": false
                  },
                  "allow_qr": {
                    "type": "boolean",
                    "description": "Enable QR payments permission",
                    "example": true
                  },
                  "allow_refund": {
                    "type": "boolean",
                    "description": "Enable refund permission",
                    "example": false
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration timestamp for the key (ISO-8601)",
                    "example": "2026-12-31T23:59:59.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API Key successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "theater_name": {
                      "type": "string",
                      "example": "Grand Cinema"
                    },
                    "api_key": {
                      "type": "string",
                      "example": "apipay_org_xxxx"
                    },
                    "status": {
                      "type": "string",
                      "example": "active"
                    },
                    "allow_invoice": {
                      "type": "boolean",
                      "example": false
                    },
                    "allow_qr": {
                      "type": "boolean",
                      "example": true
                    },
                    "allow_refund": {
                      "type": "boolean",
                      "example": false
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-12-31T23:59:59.000Z"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2026-06-25T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Admin"],
        "summary": "List All API Keys",
        "description": "Lists all generated API Keys, their status, permissions, and connection states.",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all key profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "theater_name": {
                        "type": "string",
                        "example": "Grand Cinema"
                      },
                      "api_key": {
                        "type": "string",
                        "example": "apipay_org_xxxx"
                      },
                      "status": {
                        "type": "string",
                        "example": "active"
                      },
                      "allow_invoice": {
                        "type": "boolean",
                        "example": false
                      },
                      "allow_qr": {
                        "type": "boolean",
                        "example": true
                      },
                      "allow_refund": {
                        "type": "boolean",
                        "example": false
                      },
                      "profile_id": {
                        "type": "string",
                        "example": "123456"
                      },
                      "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-12-31T23:59:59.000Z"
                      },
                      "is_connected": {
                        "type": "boolean",
                        "example": true
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2026-06-25T12:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/keys/{id}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update API Key",
        "description": "Updates permissions, name, or suspension status of a specific API key profile.",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric key ID",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Grand Cinema Updated"
                  },
                  "allow_invoice": {
                    "type": "boolean",
                    "example": true
                  },
                  "allow_qr": {
                    "type": "boolean",
                    "example": true
                  },
                  "allow_refund": {
                    "type": "boolean",
                    "example": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "suspended"
                    ],
                    "example": "suspended"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration timestamp for the key (ISO-8601). Set to null to remove expiration.",
                    "example": "2026-12-31T23:59:59.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "theater_name": {
                      "type": "string",
                      "example": "Grand Cinema Updated"
                    },
                    "api_key": {
                      "type": "string",
                      "example": "apipay_org_xxxx"
                    },
                    "status": {
                      "type": "string",
                      "example": "suspended"
                    },
                    "allow_invoice": {
                      "type": "boolean",
                      "example": true
                    },
                    "allow_qr": {
                      "type": "boolean",
                      "example": true
                    },
                    "allow_refund": {
                      "type": "boolean",
                      "example": true
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-12-31T23:59:59.000Z"
                    },
                    "is_connected": {
                      "type": "boolean",
                      "example": true
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2026-06-25T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Key profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Revoke API Key",
        "description": "Permanently deletes and revokes an API key profile.",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric key ID",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "API key successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "API Key for 'Grand Cinema' has been deleted."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Key profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}